requirements.txt
Django==5.0.6
python-dotenv==1.0.1
whitenoise==6.7.0
django-extensions==3.2.3
Pillow==10.4.0
django-allauth==0.63.6
django-crispy-forms==2.3
crispy-tailwind==1.0.3
anthropic==0.29.0
celery==5.4.0
redis==5.0.8
django-celery-results==2.5.1
django-celery-beat==2.6.0
